+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## Master
 
 * Prints a link to the comment in the build log [@orta][]
+* A Dangerfile can return a default export, and then Danger will handle the execution of that code [@orta][]
 
 ## 3.5.0 - 3.5.1